Fix: float to double in ActionTest

+++ src/test/java/fr/utc/miage/shares/ActionTest.java
@@ -93,8 +93,8 @@ class ActionTest {
         }
 
         @Override
-        public float valeur(final Jour aJour) {
-            return 0.0F;
+        public double valeur(final Jour aJour) {
+            return 0.0;
         }
     }
